+++ src/main/java/com/games/balancegameback/infra/repository/media/impl/ImageRepositoryImpl.java
@@ -5,10 +5,12 @@ import com.games.balancegameback.domain.user.Users;
 import com.games.balancegameback.infra.entity.ImagesEntity;
 import com.games.balancegameback.infra.entity.UsersEntity;
 import com.games.balancegameback.infra.repository.media.ImageJpaRepository;
+import com.games.balancegameback.infra.repository.media.MediaJpaRepository;
 import com.games.balancegameback.service.media.repository.ImageRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
 import java.util.Optional;
 
 @Repository
@@ -16,6 +18,7 @@ import java.util.Optional;
 public class ImageRepositoryImpl implements ImageRepository {
 
     private final ImageJpaRepository imageRepository;
+    private final MediaJpaRepository mediaRepository;
 
     @Override
     public Images save(Images images) {
@@ -23,19 +26,38 @@ public class ImageRepositoryImpl implements ImageRepository {
     }
 
     @Override
+    public void update(Images images) {
+        ImagesEntity entity = imageRepository.findById(images.getId()).orElseThrow();
+        entity.update(images);
+    }
+
+    @Override
+    public List<Images> findByRoomId(Long roomId) {
+        List<ImagesEntity> entityList = imageRepository.findByGamesId(roomId);
+        return entityList.stream().map(ImagesEntity::toModel).toList();
+    }
+
+    @Override
     public Images findById(Long id) {
-        Optional<ImagesEntity> entity = imageRepository.findById(id);
-        return entity.map(ImagesEntity::toModel).orElse(null);
+        return imageRepository.findById(id)
+                .map(ImagesEntity::toModel)
+                .orElseThrow(() -> new IllegalArgumentException("Image not found with id: " + id));
     }
 
     @Override
     public Images findByUsers(Users users) {
-        ImagesEntity entity = imageRepository.findByUsers(UsersEntity.from(users));
-        return entity != null ? entity.toModel() : null;
+        return Optional.ofNullable(imageRepository.findByUsers(UsersEntity.from(users)))
+                .map(ImagesEntity::toModel)
+                .orElse(null);
     }
 
     @Override
     public void delete(Long id) {
+        if (!imageRepository.existsById(id)) {
+            throw new IllegalArgumentException("Image not found with id: " + id);
+        }
+
         imageRepository.deleteById(id);
+        mediaRepository.deleteById(id);
     }
 }
